Calculation of Derivatives in Elasthyper-Materials: ================================================== In each material of the elasthyper-toolbox no longer gammas and deltas (defined by Holzapfel) are used. Now the real derivatives with respect to the principal invariants (or with respect to the modified invariants for decoupled materials) are defined in the material.
This derivatives are defined as following:
dPI(0) = \frac{\partial \Psi}{\partial I_{\boldsymbol{C}}}
ddPII_0 = \frac{\partial^2 \Psi}{\partial I_{\boldsymbol{C}}^2}
etc.
all derivatives are documented for Doxygen.

With this derivatives the gammas and deltas are calculated in the elasthyper-toolbox and the stress and constitutive-tensor is evaluated. 
-> No change at element-interface, just internal material changes!


Decoupled Elasthyper-Materials:
===============================
This pure isochoric or pure volumetric-materials are no longer calculated separate.
The (first and second) derivatives with respect to the modified invariants are converted in derivatives with respect to the principal invariants.
Therefore the simple conversions
\overline{I}_{\boldsymbol{C}} = J^{-2/3} I_{\boldsymbol{C}}, \overline{II}_{\boldsymbol{C}} = J^{-4/3} II_{\boldsymbol{C}}, J = \sqrt{III_{\boldsymbol{C}}}
are used.
With this derivatives with respect to the principal invariants, the gammas and deltas and therefore the stress and constitutive-tensor is calculated.
-> No change at element-interface, just internal material changes!


Calculate Strain Energy:
========================
Every material of the elasthyper-Toolbox can now calculate it's own strain energy.


Material coupaaagasser and coupHUdependentneohooke:
===================================================
Added StrainEnergyFunction for Documentation.
Renamed back to isovolaaagasser and respectively isovolHUdependentneohooke, as this materials contains both volumetric and isochoric parts, but acts physical as an decoupled material.
Set flag to isomod and use a decoupled formulation.


plastic Elast Hyper (by Alexander S.)
===============================
- adapted to new evaluation of elastic summands
- now plasticelasthyper is derived from elasthyper
  to reduce code redundancy
- adapted plastic TSI (with semi-smooth Newton) accordingly


Testcases
=========
- adapted slightly different results of shell-elements in solid_ele_vs_mat-testcase
- adapted slightly different pressure result in mfsi_ow3D_mtr_drt.dat (thanks to Matthias for the help!)
- changed names of coupaaagasser and coupHUdependentneohooke in calcified_zugstab and elasthyper_IsotropicMAT_stress_free_reference.dat


Documentation:
=============
Added new derivatives and conversion of derivatives with respect to modified invariants in derivatives with respect to principal invariants to doxygen.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20599 060e22d8-9d1c-0410-868a-e3a950d63d95