Refactoring of DRT::Element ===========================
Previously, DRT::Element was general enough to allow for both standard elements (that we read from a file) and 'child' elements that are associated to a parent element, i.e., those that return non-NULL ParentElement() or ParentMasterElement(). Such elements are typically boundary elements, but may also be Mortar elements or interior face elements. We used access to ParentElement() in many places without actually asserting that the respective element supports this feature.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20799 060e22d8-9d1c-0410-868a-e3a950d63d95