Continuation of commit message from r20799 ==========================================
In the new design, an base element class DRT::FaceElement for elements that are associated to a parent have been introduced. This makes access to parents explicit and requires dynamic_cast at a few points, but is safer because it clearly states the intent of the element. This also frees 24 bytes per element for all the standard elements that do not need to store pointers to parent master and slave as well as their numbers any more.

Moreover, the diamond structure of the element MortarElementGeoDecoupl, deriving from both MORTAR::MortarElement and DRT::MESHFREE::Cell, has been removed in favor of making the element base of DRT::MESHFREE::Cell a template. Multiple (diamond) inheritance is nice on the algorithm level but it is cleaner to avoid it on the element level because we want to have slim and simple data structures there.

(I pressed the wrong button when writing the commit message of r20799, sorry for that.)

With this commit, I introduced the local transformation map into FaceElement that is necessary for interior faces. This avoids implementing the same feature several times in derived classes (Acou, Combust3, Fluid). On boundary elements, the map is not filled, keeping the memory overhead limited for all the elements that do not need it.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20802 060e22d8-9d1c-0410-868a-e3a950d63d95