First step to make HDG elements work with periodic boundary conditions: Fix AdjustFaceOrientation
AdjustFaceOrientation now finds the renumbering on quadrature points via the LocalTrafoMap that gets stored within the face elements. In MPI, this list needs to be communicated. On ghost elements, we need to make sure the master/slave relation of the element is correct, which leads to the situation that an element might only be a slave element but not a master element.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20804 060e22d8-9d1c-0410-868a-e3a950d63d95