let surface current evaluation make use of parallel layout: ----------------------------------------------------------
the evaluation is inherently not a very nice method for computers (convolution, and several of them). So far all work was left to processor 0 which takes infeasibly long for already moderately sized problems. Now, with some communication of data involved, the convolutions workload can be shared among all processors which gives a nice speedup for the problem i tested with. But this method will stay slow by definition.

adapted testcase

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21293 060e22d8-9d1c-0410-868a-e3a950d63d95