==| CUT - Fixes for Creating Cut Lines |==
Treatment of special cases for creating cut lines was not correct! 
Just adding this new cases did not work, as the construction of cut lines is a bit confusing!

- New Implementation for creating cut lines in a more straight forward way!
--> Special Case with more than two cut_points(arising from tolerances) from intersection of two planes is fixed!

(there are still problems with the pointgraph for creating point cycles and finally facets ..., which is the next step after creating lines)

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21313 060e22d8-9d1c-0410-868a-e3a950d63d95