Generate build configuration for Fedora 22.
This includes a new config file fc22_Q12015.config that sets up compilation of BACI with the system-provided gcc 5.1 and the support libraries in /lnm/lib/Q1_2015/fc22. For this configuration, BACI is compiled in C++11 mode. For the time being, MueLu is disabled with this configuration because the interface in 12.3 is different from the old interfaces. I will switch to a new version once we have moved our computers to the new Fedora system.

Finally, I also added CFLAGS (separate from CXXFLAGS) to all config files. C flags must not contain the -std=c++11 switch which is added to C++ compilation (obviously), thus we need to have different flags.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21318 060e22d8-9d1c-0410-868a-e3a950d63d95