Fix failing test cases due to r21361.
In my previous commit I immediately released all instances of ScaTraEleCalc* and friends when the first instance should be deleted by Done(). Of course, this is the wrong thing to do since the global problem wants to delete all singletons individually, i.e., calling Done() on the instances separately. Therefore, the Done() call should only delete the 'this' pointer, which is solved by passing the 'this' pointer from Done() to Instance(). The pointer replaces the old 'bool create': If no pointer is given (default case), we create a new instance, otherwise we delete the object in question.

This patch looks big since it applies the correct mechanism to all ScaTra singletons, but the relevant code are just a few lines in the Instance() and Done() methods, the rest is just repetitions of the same concept.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21363 060e22d8-9d1c-0410-868a-e3a950d63d95