Enable MueLu for Q1_2015
Based on a preliminary patch of Tobias Wiesner, I have adapted MueLu for the new Trilinos Q12015. In order to avoid lots of ifdefs, I have disabled MueLu on Q3_2014 (the experimental track on the old FC16) and it is now only available for FC22 and Q12015. For kaiser, there will be a new Q12015 config file in a few days with MueLu enabled when we are done with compiling the new Trilinos.

If you were running Q32014 on FC16 before, you will get lots of compile errors. To fix them, you need to re-run do-configure that disables MueLu on that target.

Likewise, if you would like to use MueLu and have already run do-configure on FC22 before this patch, re-run do-configure --experimental to enable MueLu.

This patch has become necessary due to changes in the MueLu/Tpetra/Kokkos interfaces that are not within our control.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21371 060e22d8-9d1c-0410-868a-e3a950d63d95