========== Photoacoustic image reconstruction with respect to acoustic material parameters
Up to now, it was only possible to reconstruct the optical material properties. From now on, one can also reconstruct the density and the speed of sound of the acoustical discretization. This is computationally intense, since the computation of the gradient combines solutions of forward and adjoint solutions at each time step. And one cannot store all solution vectors of the forward problem, this would require overmuch storage. Therefore, I implemented a simple checkpointing routine: in the forward run, restart steps are written each RESTARTEVRY step. In the adjoint run (which is backwards in time), a restart from the forward problem is read, RESTARTEVRY forward steps are performed and RESTARTEVRY solution vectors are stored; then the correspondent adjoint steps are executed and the gradient contributions are calculated. Doing so, one only has to store the forward solution for RESTARTEVRY steps but the adjoint integration takes twice as long as the forward integration.
- implemented a function IntermediateIntegrate() in the implicit scheme and analogous intermediate_integrate() in the explicit WaveEquationProblem which performs the intermediate forward steps
  - reads correspondent restart
  - initializes the elements
  - performs RESTARTEVRY time steps
  - resets everything such that the adjoint run can go on
- the elements have a initialization routine and a gradient contribution computation
- in the gradient computation forward and adjoint states are combined and written to the elements
- elements have GetGradient(), AddToGradient() and InitGradient() functions
- the inverse analysis class has now two MatParManager (scatra_matman_ and acou_matman_), two regularizations and so on
- adapted input parameters accordingly
- implemented an AcouMatParManagerPerElement and Uniform as derived class of corresponden MatParManager to accomodate the special needs of acoustic materials


========== Photoacoustic image reconstruction with respect to material concentrations

The photoacoustic inverse problem is ill-conditioned and (especially when one does not only optimize optical properties but also acoustical properties) the solution is not unique. To address these two issues and to make use of prior knowledge on material, I implemented a "segmentation". The user defines possible materials (for example if I image a mouse torso, I define materials for heart, liver, lung, bone, ...; if I image a human finger, I define materials for skin, bone, muscle (such values can be found in the literature) with speed of sound, density, absorption coefficient and diffusion coefficient) and then the optimization only optimizes the concentrations of these materials and no longer all four material parameters independently. Also, to keep the concentrations between 0 and 1, they are again parametrized just like the METAPARAMS in standard inverse analysis using the arctan for transformation from (-inf,inf) to (0,1).
- input parameter SEGMENATION boolean to indicate if one wants to run the segmentation
- input parameter SEGMENTATIONMATS as string with a file name where the materials are defined
- the acoustical inverse analysis has a k_objgrad_ etc. just as the scatra and the acoustic field
- the acoustical inverse analysis has a data type Segmentation which handles the initialization of the concentrations and the calculation of the gradient


========== Photoacoustic image reconstruction with explicit time integration

The photoacoustic image reconstruction now works with explicit time integration for the acoustic problem. The explicit time integration is implemented using DEAL, hence all the framework had to be adapted such that the additional information for the inverse problem is brought to the time integrator.
- the time loop of the WaveEquationProblem now has the history as input parameter as well as the output
- added the functionality for the intermediate time integration (for optimization with respect to acoustical parameters)
  - stored_forward_solutions
  - intermediate integrate (analogue to the implicit one)
- the WaveEquationOperationBase now has functions and functionality to compute and write gradient contributions for the gradients with respect to density and speed of sound
- added some comments


========== Delete of all implicit acoustic time integrators except for the implicit Euler

Yes, sometimes it is useful to delete something. Now that the inverse analysis runs with the explicit time integration, I will keep only the implicit Euler as reference and for tests. Maybe the DIRKs (diagonally implicit Runge Kutta schemes) will return but they would need rework anyhow and right now I do not need them. The trend goes to explicit time integration!
- some changes in the base class AcouTimeInt and the base class for implicit schemes AcouImplicitTimeInt
- several simplifications in the element evaluation routines


========== Misc

- fixed the bug Karl's Hiwi found
- rewrote the line search routine, now as a class PATLineSearch and not longer somewhere in ACOU::InvAnalysis
  - three different run types: RunSuccessive() (for successive optical and acoustical update), RunSimultaneous() (for simultaneous update or in case only one field is optimized) and RunSegmentation() (as described above)
- adapted the test cases for photoacoustic image reconstruction
- added test cases for the explicit time integration with deal (they will be enabled as soon as more than one testing machine is on Fedora 22)
- updated the config files to enable deal and the correspondent test cases when configuring with the option --useDEAL


NOTE: This commit is from Fedora 22.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21372 060e22d8-9d1c-0410-868a-e3a950d63d95