Disable ACML on Q1_2015 for FC22.
Due to a bug that manifests itself at the initialization of global variables (before main() is entered), ACML must be disabled.

With this commit, the --experimental track is hopefully fully functional (both on kaiser and FC22), providing the following features:
- Trilinos 12.3 (version from August 27, 2015), MueLu enabled
- gcc 5.1 on FC22, gcc 4.8 on Kaiser
- OpenBLAS v0.2.14 for high-performance BLAS/LAPACK implementations on FC 22
- ACML on kaiser
- C++11 enabled (-std=c++11) (forced by Trilinos)

Please note that using C++11-isms in code is still not allowed because the non-experimental code is still on C++03. Be careful when developing on the --experimental baci.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21377 060e22d8-9d1c-0410-868a-e3a950d63d95