Re-use the matrix structure upon SparseMatrix::Zero() for increased performance.
Previously, a call to Zero() on a LINALG::SparseMatrix (the most common variant that is constructed without a graph) would clear the previously generated non-zero pattern (graph) of the sparse matrix. With this change, a matrix that is Zero()'d only writes zeros into the entries associated to its sparsity pattern (without inserting new elements) in case
(a) the matrix was filled
(b) the matrix was contructed with explicitdirichlet_ set to false (explicitdirichlet_ creates a new matrix in ApplyDirichlet which is not compatible with assembly in a later time step).

To make use of the less frequent memory allocation (and expensive calls that transform between global and local indices in lung computations in particular), I adapted the matrix construction in the structural time integrator to create matrices with explicitdirichlet_ set to false.

Even though I checked that all test cases still work (there are two known problems with nestedpar_copydatfile-nestedPar_CopyDat-p2 and sohex8_easfull_cooks_nl-p2 to be addressed elsewhere), this change is sensitive and needs your attention. If you have sparse matrices whose sparsity pattern changes over time (e.g. some mortar/contact problems, XFEM, ...), you need to use LINALG::SparseMatrix::Reset() when you generate a new matrix. If the sparsity structure of your matrix is unchanged, use LINALG::SparseMatrix::Zero() and (hopefully) profit from increased performance. If unsure, talk to me. If there are problems that I overlooked, talk to me.

The general rule should be that we try to use Zero() as much as possible, in particular in the primary solvers of fluid and structure.

For TSI problems, I changed some SparseMatrix::Zero() calls into SparseMatrix::Reset() calls to keep some volume mortar problems working.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21554 060e22d8-9d1c-0410-868a-e3a950d63d95