Configuration file for new cluster named 'bruteforce'.
The current configuration uses gcc-5.3 (same major version as on FC22) and OpenMPI 1.10 for compilation.

I will also add a configuration script with the Intel compiler v16.0 before the cluster is opened for general usage.

The location of libraries is similar to the cluster named 'deep/thought', with the generic code located in '/lnm' and the version-specific code (Trilinos, deal.II) in '/lnm/Q1_2015'.

Result of minimal ctest on new cluster:
$ ctest -L minimal
[...]

100% tests passed, 0 tests failed out of 24




git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21731 060e22d8-9d1c-0410-868a-e3a950d63d95