Safety check for computing adapted relative tolerances.
When using the option STRUCTURAL DYNAMICS -> ADAPTCONV true, the inproper
choice of parameter STRUCTURAL DYNAMICS -> ADAPTCONV_BETTER can lead to
relative tolerances >1 for the linear solver. This would lead to 
finish the solver immediately irrespective the current residual.

Therefore the adapted tolerance is checked and set to 1 if its
value is bigger than 1. Additionally a WARNING expression will
show up.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21790 060e22d8-9d1c-0410-868a-e3a950d63d95