Fix compilation with clang's libc++
Some XFEM classes have std::map<XFEM::DofKey, XFEM::DofGID> that gets copy-assigned. The problem is that XFEM::DofKey does not provide a copy assignment operator. gcc's implementation of the C++ library seems to implement this by only copy constructors, which is not the case for the native libc++ of the clang compiler. To avoid compiler-specific behavior, I manually add all entries to the map which is fine with copy constructors.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21860 060e22d8-9d1c-0410-868a-e3a950d63d95