New Point Coupling Conditions (and corresponding DoF handling) in BACI
Some applications (typically in structural / solid mechanics) require the coupling of certain DoFs of two or more nodes at the same geometrical position, while certain other DoFs of those nodes shall remain independent (e.g. joints and hinges in frames). While it is very easy to couple all(!) DoFs of several nodes at the same geometrical position (by simply merging the nodes into one node), things are more complicated if only certain DoFs are to be coupled. While it would always be possible to introduce this coupling as a Dirichlet condition / Multipoint Constraint into the final system of equations, we have decided to implement this at a more fundamental level by changing the assigment of DoFs according to the existing coupling conditions. Thus, if a point coupling condition is introduced for a set of nodes, the DoFs to be coupled are identified and the same(!) DoFs are then assigned to all participating nodes, while the remaining uncoupled DoFs are created and assigned independently for each node. This required some changes in the way nodal DoFs are assigned and handled in BACI. However, after the initial DoF handling, the nice thing about this approach is that nothing needs to be done anymore at the system matrix level because the coupling is inherently included in the DoF-maps. If you think of a web-like frame structure with many joints and hinges, this also means that the global system size is drastically reduced as compared to a Dirichlet type handling of such constraints.

Included in this commit:
- new point coupling condition - e.g. for joints / hinges in structural mechanics
- no interference (hopefully) with element or face DoFs
- DofSet class is now able to handle repeated assignment of DoFs to more than one node
- DofSet class is now tracking and storing not only the first DoF ID of a node but all DoF IDs of a node

NOT included in this commit:
- support for derived DofSet classed that overload AssignDegreesOfFreedom (e.g. MortarDofSet, PotentialDofSet)
- support for special DofSet stuff (e.g. TransparentDofSet, Proxies...)
- support for bandwidth optimization (#define BW_OPT), which is currently however not used anyway

Input file usage:
- this is how the input file would look like for a Y-junction of three Reissner beam elements with a pin joint ("Momentengelenk")

------------------------------------DESIGN POINT COUPLING CONDITIONS
DPOINT 1
E 1 - NUMDOF 6 ONOFF 1 1 1 0 0 0
-------------------------------------------------DNODE-NODE TOPOLOGY
NODE       17 DNODE 1
NODE       42 DNODE 1
NODE       99 DNODE 1


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21874 060e22d8-9d1c-0410-868a-e3a950d63d95