As announced three weeks ago, test case redundancies through serial as well as parallel testing are reduced: -> Deleted all serial tests which were tested on 1 processor and on several processors (except for those for which I got an email).
If you are not happy with the delete of some of your serial tests, think about it if you really need the additional serial check and if yes, just (re-)add the corresponding line in TestingFramework.cmake. But keep in mind that the coverage of the code through this commit should not be affected significantly. Also, if you have a closer look at your tests anyhow, please consider Karl's email (see  below).



Dear colleagues,

if I may add to this:

A HiWi of mine wrote and ran a script over all testcases, which
determines their sensitivity (i.e. usefullness). For this most input
parameters were varied until the testcase fails and the .dat-file was
annotated with the results.
While cleaning up the tests please also take a look at the annotated
tests in 

/lnm/network/annotated_input/ 

and check whether the test is

1. usefull/needed
2. sensitive to the parameters that are to be tested
3. all defined parameters are used

vg
Karl

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21910 060e22d8-9d1c-0410-868a-e3a950d63d95