fixed contact module: =============
For the contact module we extract files from baci and modify them with patches such that it works as seperate module. Here, the rule of thumb is that corresponding files should include only headers from the mortar/contact folders itself. This is necessary to keep the contact module functional and maintainable on a long term basis. 

If you see no alternative than including files from other folders please contact me and we will find a solution.

specific modifications in new patches:
- removed nitsche contact formulation since it requires so_base.H (It is well-known to contact group members that this is not allowed!)
- removed poro related functions


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21913 060e22d8-9d1c-0410-868a-e3a950d63d95