Signal handling for Floating Point Exceptions
Whereas OpenMPI issues a more or less readable stack trace through MPI_Abort in case of floating point exceptions, the Intel MPI environment doesn't and just exits with the very informative 'Your process exited with signal floating point exception'. So I installed a signal handle for SIGFPE. If compiled with TRAP_FE, it raises a dserror in case of SIGFPE and you get the nicely readable stack trace from the catch of runtime_errors in ntam.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21985 060e22d8-9d1c-0410-868a-e3a950d63d95