Polyconvexity check ===================
It is now possible to check the polyconvexity of isotropic hyperelastic materials (dependent on principal or modified invariants).

Definition of Polyconvexity:
\Psi(F) ist polyconvex, if there exists a function
\Psi(F) = P(F,CofF,detF)
and this function is convex for all points X.

This is true if the second Frèchet derivative of (F, CofF, detF) is positive (semi-)definit.

The second Frèchet derivative according to eq. (5.31) of Vera Ebbing - Phd-thesis (p. 79) is calculated.
This calculation is valid for all hyperelastic (dependent on principal or modified invariants) materials, as the partial derivatives of the SEF are used for the calculation of eq. (5.31).

To check the polyconvexity your input-line should read:
MAT 1 MAT_ElastHyper NUMMAT 1 MATIDS 2 DENS 0.0002 POLYCONVEX 1 
By default polyconvexity is not checked, parameter POLYCONVEX is optional.

The check of anisotropic, viscoelasitc or plastic materials is not possible with this function.

Tested with help of the non-polyconvex elast_isotestmaterial and elast_volsussmannbathe, as they contain all possible derivatives of the SEF.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22005 060e22d8-9d1c-0410-868a-e3a950d63d95