Removed the return of "const double" and "const int" returns in two files. These returns give rise to warnings on the bruteforce cluster when building baci. It is also bad programming practice as the "const double/int" return can not be enforce.
Having a look at the two possible cases:

1) Case of int/double/pass by value return
Assume the function is defined as,

const int GetNumDof() {return numdof_;}

no error is given if the function is called with 

int numdofs =  foo.GetNumDof();

i.e. the const before the int is not enforced. However 

int GetNumDof() const {return numdof_;}

is resonable as this ensures that no member variables are changed inside this function unless they are set as mutable.


2) Case of pointer/pass by reference return
For a pointer a const return is sound i.e.

const std::vector<double> &GetVector() const { return vector_; }

is good programming practice.


Too make this commit I had to assign a level to the changed files. So I set them to level 3, would the Maintainer of these files change them to an appropriate level when they have the time?

These changes have been discussed with Martin K.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22042 060e22d8-9d1c-0410-868a-e3a950d63d95