_____________________________________________    Fixes inconsistency in DBC Handling _____________________________________________
Patch by Andreas Rauch

Feature:

Some of us have been using a certain feature when applying Dirichlet BCs.
It may be advantageous to define DBCs on a nodeset in your input file and
release them afterwards on a subset of higher priority.

e.g. You may define a volume dirichlet condition on an irregular
     object. However, you want to release the surface nodes from
     this condition. It is difficult and onerous to define the
     nodeset for the inner nodes in cubit or trelis. It is much
     easier to achieve this in the following manner:

    1) Define the volume dirichlet condition:
    ---------DESIGN VOL DIRICH CONDITIONS
    DVOL   1
    // example_volume
    E 1 - NUMDOF 3 ONOFF 1 1 1 VAL 0.0 0.0 0.0 CURVE none none none
FUNCT 0 0 0

    2) Release the outer surface
   ---------DESIGN SURF DIRICH CONDITIONS
   DSURF  1
   // example_surface
   E 1 - NUMDOF 6 ONOFF 0 0 0 VAL 0.0 0.0 0.0 CURVE none none none FUNCT
0 0 0

    Now only the inner nodes of the example_volume have DBCs, since
    surface dirichlet conditions are supposed to overwrite the volume
    conditions. This has been used in some applications by some co-
    workers, and seemed to work perfectly well. However, there has
    been a hidden bug that may influence your results.
   

Problem Description:
      
Even though the Dirichlet application is supposed to follow a priority
rule, stating that point DBCs overwrite line DBCs, which again overwrite
surface DBCs, which finally overwrite volume DBCs, there was an
inconsistency to this rule when applying the feature explained above.

By first calling :

    DoDirichletCondition(DRT::Condition::VolumeDirichlet,usetime,time,
                         systemvector,systemvectord,systemvectordd,
                         toggle,dbcgids) ,

we already assigned the values specified in the input file to the
systemvectors! Next, we called

    DoDirichletCondition(DRT::Condition::SurfaceDirichlet,usetime,time,
                         systemvector,systemvectord,systemvectordd,
                         toggle,dbcgids).

Unfortunately, for the example given above the SURF DIRICH dofs were now
erased from the set 'dbcgids', which is fine, but the VOL DIRICH values
had already erased necessary content in the systemvectors!


Solution:

In EvaluateDirichlet( ... ):

The idea is to gather the information on the Dirichlet dofs first and only
apply them in a second step.

Now we call (starting from lowest and proceeding to highest priority):

\code
    if (fool->first != "Dirichlet") continue;
    if (fool->second->Type() != DRT::Condition::VolumeDirichlet) continue;
    ReadDirichletCondition(*(fool->second),toggleaux,dbcgids);
\endcode

    ... Surface -> Line -> Point

At this stage, we only fill the toggle vector and the dbcgids map if
requested.

After this code has run for the entity with highest priority, i.e. for
DRT::Condition::PointDirichlet, the toggle contains all dof gids for
which a DBC should really be applied.

Then we can call:

\code
  // Now, as we know from the toggle vector which dofs actually have
  // Dirichlet BCs, we can assign the values to the system vectors.
  // Assign VolumeDirichlet dbcgids first
  for (fool=condition_.begin(); fool!=condition_.end(); ++fool)
  {
    if (fool->first != "Dirichlet") continue;
    if (fool->second->Type() != DRT::Condition::VolumeDirichlet) continue;
    DoDirichletCondition(*(fool->second),usetime,time,
                         systemvector,systemvectord,systemvectordd,
                         toggleaux);
\endcode

    ... Surface -> Line -> Point

for each Dirichlet type. This actually assigns the values to the
systemvectors.

Now, only values that are contained in the dbcmap are touched in the
system vectors.
This was not the case before.
 
Since DBC application had caused trouble earlier, Andreas Rauch will give
a short mini-CMGM on this topic soon.
From testing the corrected DBC application he got the impression that some
of us might have applied DBCs wrongly in a systematic fashion. This is now,
and has always been wrong! Only now it may be elucidated.

Affected Maintainers:

Anh-Tu Vuong
Matthias Mayr
Benjamin Krank
Ursula Rasthofer
Volker Gravemeier



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22109 060e22d8-9d1c-0410-868a-e3a950d63d95