Patch by Andreas Rauch.
> BUGFIX OF REVISION 22486
> __________________________________________________________________________
o Fixes failing xf* tests.
o The purpose of the patch for revision 22486 was to make the parallel distribution
  of conditions and derived discretizations more determinisitc, i.e., every geometry
  element is reliably owned by the underlying volume element.
o It turned out that now it is possible that a surface element is owned by a processor
  which ghosts all the element's nodes.
o This happens if a volume element happens to have only ghosted nodes at one face which
  is supposed to wear a condition.
o The reason why this is a problem, is the fact that we loop over conditiond row nodes,
  and only over surfaces which have these row nodes attached, in order to decide if a surface
  is to be created or not.
o Unfortunately now we might miss a face unless we loop over conditioned column nodes.
o Looping over conditioned column nodes fixes this issue with minor increase of computational
  effort.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22507 060e22d8-9d1c-0410-868a-e3a950d63d95