==| Fix testcase xff_beltrami_TesTes_NIT_embedded_evp |==
- Loose the pressure error result check from 1e-8 to 1e-7

the reason for this is, that this testcase solves a generalized eigenvalue problem to determine the h (Scaling between volume and interface viscous term). h is very sensitive to the matrixes, which are input for the generalized eigenvalue problem and the testcase is very sensitive to h.
--> pressure error changes more than 1e-8 due to computational accuracy, 1e-7 seem to be fine.

Thx to Alex S. for analyzing

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22575 060e22d8-9d1c-0410-868a-e3a950d63d95