Much of the following is based on the work by HiWi Stefan Legat.
Added a unit testing framework:
===============================

What is unit testing?
https://en.wikipedia.org/wiki/Unit_testing

Why would I want to use unit tests in BACI?
The existing tests test the overall interaction of BACI parts nicely, but in the end you judge whether all 850000 LOC in BAC are correct by a handful of doubles. This leaves you two options: Testing just some parts and hoping the rest will be fine as well or writing millions of test cases, which test every possible combination of features and take an prohibitively long time to run (right now we are testing just small parts _and_ it takes forever..).
Another problem is the sensitivity of the tests. For example it is very difficult to detect small errors in the linearization since all simulations iterate until the tolerance is fulfilled and whether it took 2 or 15 iterations is not tested. Or there may be a bug in a stabilization term in a corner of the domain which will most likely not trigger the 1e-8 error tolerance of the x-velocity on the other side of the domain.
This is where unit tests come in handy. It gives you the possibility to test methods and classes directly and without the indirection of placing it in an entire simulation. This makes tests fast, sensitive, since the results are checked directly and allows for pathological cases to be tested as well.

How are the unit tests implemented in BACI?
Everything is encapsulated in the Unittests folder, which is integrated into the build process by the UnittestingFramework.cmake. The testing framework that operates in the background is "CxxTest" (http://cxxtest.com) and handles test discovery, testing and error handling for you.
The directory structure inside the Unittests folder is analog the BACI directory structure and all unit tests are named according to the file/class they are testing.
Additionally there is a folder called stubs, which holds stubs of BACI classes which are needed by tests, but for which the actual classes cannot be used (perhaps due to file IO, etc.). Unfortunately the attempts at automated stub generation had to be abandoned. :(

How do I run the unit tests?
Choose one of the following:
1: Unitests/unittest
2: ctest -R unittest
3: ctest -L minimal
4: make test

How do I add a unit test?
Lets say you would like to test a class in a file src/drt_bring/drt_coffee.H. Then simply copy an existing unit test file, such as unit_so_hex8.H, to Unittests/src/drt_bring/unit_drt_coffee.H and add the new unit test to the framework by appending it to the list in Unittests/src/drt_bring/CMakeLists.txt. Now adapt the testcases for your class. Build and run the tests with "make unittest -j6 && ctest -R unittest".
Should the folder for your test not exist, copy another test folder, remove the contained unit tests and adapt the Unittests/src/CMakeLists.txt accordingly.
If you managed to do the above, you should be able to figure out how to add stubs by yourself.

How do I make the best use of unit tests?
Read "Clean Code" by Robert C. Martin.

Are there some rules for using unit tests in BACI?
No, not yet, but here are some general guidelines:
- use classical testing (see http://martinfowler.com/articles/mocksArentStubs.html for a very good clarification on the two main unit testing approaches)
- test only public methods
- use original classes/objects as often as possible
- if that is not possible, write a stub
- for now all tests are run using a single MPI thread
- tests have to be independent of one another
- write many tests (they cost almost nothing)
- test corner cases
- test failures

Other minor changes:
- add -D_GLIBCXX_USE_CXX11_ABI=0 to the compiler options, which enables gcc >= 5 to compile and link BACI against Trilinos compiled with older compilers
- adapted adapt_cproject.py to include the cmake generated headers as well

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22586 060e22d8-9d1c-0410-868a-e3a950d63d95