Introduced interaction between airways and parenchyma for 0D lung models. Now a compliant airway senses the external pressure of the surrounding tissue (i.e., of the nearest 0D acinus) and expansion/compression of an airway is governed by the difference between airway inner and external pressure.
To include this interaction in 0D lung computations, set the flag COMPAWACINTER to yes in the --REDUCED DIMENSIONAL AIRWAY DYNAMIC section of the input file.

This is work done by Anna-Katharina Rettinger in her Bachelor thesis where more details can be found.

Also added a testcase for airway-acinus interaction code.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22624 060e22d8-9d1c-0410-868a-e3a950d63d95