* Changed gauss-point weight calculation for TRI's coming from cut. * Adapted test-cases affected by this change (3-xfem two phase flow cases (press tol 1e-14 -> 1e-13), the altgeo case has same result only beauty) * Added a test case which was not in the TestingFramework.cmake but in the repository * Added Christoph as maintainer for cut_boundarycell.H with level 3. * Minor: surpressed output of Get_ViscPenalty_Stabfac for CouplingBase in XFEM
As the test case, xfsi_comp_struct_fsi_2D_mono_BGS2x2_infnormscal, has been failing for quite some time, I looked into it.
The issue stemmed from calculating the gauss point weights using the metric tensor for small/ill-conditioned TRI's. This could lead to negative determinants, thus having simulations fail with floating point error, which was the case for the previously mentioned test case.

The solution is to use an area calculation and assign the gp-weights this way. Furthermore, as the choice of which TRI's are kept comes from the same routine, this way of assigning the weights will be more coherent.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22664 060e22d8-9d1c-0410-868a-e3a950d63d95